gltf

+++ src/pt_gfx_mesh_base_gltf_lex_yacc.c
@@ -68,7 +68,7 @@
 union YYSTYPE
 {
     int m_token_numberint;
-    int m_token_numberfloat;
+    float m_token_numberfloat;
 
     int m_scene_index;
     int m_node_index;
